Get rid of windows specific freopen
